Possibilidade e escolher o tipo de inteiro com int ou long.

+++ include/Adjcency.h
@@ -3,12 +3,13 @@
   #include<Memoria.h>
   #include<Mystdbool.h>
   #include<Mesh.h>
-  void viz(Memoria *m ,long *el   ,long *nelcon
-          ,short *nViz,short *nen ,long nnode
-          ,long numel ,short maxNo,short maxViz);
-  void adj2d(long *el   ,long *nodcon ,long *nelcon
-            ,short *nen ,long numel   , long nnode 
-            ,short maxNo, short maxViz,long *nEdeg);
+  #include<Define.h>
+  void viz(Memoria *m ,INT *el   ,INT *nelcon
+          ,short *nViz,short *nen ,INT nnode
+          ,INT numel ,short maxNo,short maxViz);
+  void adj2d(INT *el   ,INT *nodcon ,INT *nelcon
+            ,short *nen ,INT numel   , INT nnode 
+            ,short maxNo, short maxViz,INT *nEdeg);
 
 #endif/*_MESH_*/
