+++ include/File.h
@@ -5,19 +5,20 @@
   #include<Mystdbool.h>
   #include<math.h>
   #include<string.h>
-  #define MAX_STR_LEN_IN 129 /*tamanho do arquivo de entrada maximo*/
+  #include<Define.h>
+  #define MAX_STR_LEN_IN    129 /*tamanho do arquivo de entrada maximo*/
   #define MAX_STR_LEN_SUFIXO 80  /*tamanho do arquivo de entrada maximo*/
-  #define SIZEMAX 200 /*tamanho do arquivo de saida maximo*/
-  #define MAX_STR_NUMBER 65 /*long integer*/
-  #define MAX_EXT        20 /*numero maximo da extencao*/ 
-  #define WORD_SIZE   120
-  #define LINE_SIZE   120
-  #define MAX_LINE  200
+  #define SIZEMAX           200 /*tamanho do arquivo de saida maximo*/
+  #define MAX_STR_NUMBER     65 /*INT integer*/
+  #define MAX_EXT            20 /*numero maximo da extencao*/ 
+  #define WORD_SIZE         120
+  #define LINE_SIZE         120
+  #define MAX_LINE          200
 /*... uso na depuração de erros*/
 /*...................................................................*/  
-  void  iota(long,char*);
+  void  iota(INT,char*);
   FILE* openFile(char *,char *);
-  void  fName(char*,long,int,char**);
+  void  fName(char*,INT,int,char**);
   void  readMacro(FILE*,char*,bool);
   void  clearLine(char *);
   int   rl(FILE *,char *);
