+++ include/ReadFile.h
@@ -14,16 +14,15 @@
   #include<File.h>
   #include<Mystdbool.h>
   #include<Memoria.h>
-  void parametros(long int*nnode,long int *nel,short *maxno,short *ndm
-                 ,short   *numat,short    *ndf,FILE  *file);
+  void parametros(INT   *nnode,INT *nel   ,short *maxno,short *ndm
+                 ,short *numat,short  *ndf,FILE  *file);
   void readFileFvMesh(Memoria *m,Mesh *mesh, FILE *file);
-  void readVfCoor(double *x,long int nn, short int ndm,FILE *file);
-  void readVfElmt(long int *el    ,short int *mat ,short int *nen
-                 ,short int *ty  ,long int nel   ,short int maxno
+  void readVfCoor(double *x,INT nn, short ndm,FILE *file);
+  void readVfElmt(INT *el    ,short *mat ,short *nen
+                 ,short *ty  ,INT  nel   ,short maxno
                  ,FILE *file);
-  void readVfRes(short int *id,long int numel,short int maxno
-                ,char *str    ,FILE *file);
-  void readVfSource(double *f ,long int numel,short int maxno
+  void readVfRes(short *id,INT numel,short maxno,char *str,FILE *file);
+  void readVfSource(double *f ,INT numel,short maxno
                    ,char *str ,FILE *file);
   void config(bool *bvtk,Reord *reord,FILE* f);
 #endif
